TodoApp/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pydantic>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pymysql>=1.1
